requirements.txt
awscli==1.44.81
blinker==1.9.0
boto3==1.42.91
botocore==1.42.91
cffi==2.0.0
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
docutils==0.19
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
MarkupSafe==3.0.3
pyasn1==0.6.3
pycparser==3.0
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
rsa==4.7.2
s3transfer==0.16.0
six==1.17.0
urllib3==2.6.3
watchtower==3.4.0
Werkzeug==3.1.8
